Reference frame modification
Change to the equatorial reference system for all celestial bodies. When a new save is opened, the relative positions of the celestial bodies are the same as in 1950.1.1 in RealSolarSystem Mod. The purpose of this is to be compatible with some RSS-related Mods, such as Charon's RSS-Origin. At the same time, in order to adapt to Principia soon. This change causes the atmospheres of all celestial bodies to be modified, but don't worry, they are unchanged, only the temperature is modified